requirements.txt
cv2
numpy~=1.18.1
torch>=1.0
torchvision~=0.2.1
matplotlib~=3.1.1
tensorflow
tensorboard
terminaltables~=3.1.0
pillow==6.1
tqdm~=4.38.0
opencv
shapely~=1.6.4
colorama
visdom~=0.1.8.9
yaml~=0.1.7
skimage
pyyaml~=5.1.2
scikit-image~=0.15.0
scipy~=1.3.2
pandas~=0.25.3
pycocotools~=2.0.0
scikit-learn~=0.22.1